added dashboard component, added routing, refactored login and gegister forms
