+++ listwait-app/src/components/Create.js
@@ -1,0 +1,79 @@
+import React, { Component } from 'react'
+import Header from '../components/Header';
+import Footer from '../components/Footer';
+import Register from '../components/Register';
+import Login from '../components/Login';
+
+export default class Create extends Component {
+    constructor(props) {
+        super (props);
+        this.state = {
+          isLoggingActive: true
+        };
+      }
+
+    componentDidMount() {
+        this.rightSide.classList.add("right");
+      }
+    
+      changeState() {
+        const { isLogginActive } = this.state;
+    
+        if (isLogginActive) {
+          this.rightSide.classList.remove("right");
+          this.rightSide.classList.add("left");
+        } else {
+          this.rightSide.classList.remove("left");
+          this.rightSide.classList.add("right");
+        }
+        this.setState(prevState => ({ isLogginActive: !prevState.isLogginActive }));
+      }
+
+    
+      render() {
+        const { isLogginActive } = this.state;
+        const current = isLogginActive ? "Register" : "Login";
+        const currentActive = isLogginActive ? "login" : "register";
+        return (
+          <>
+          <Header />
+          <div className="App">
+            <div className="login">
+              <div className="container" ref={ref => (this.container = ref)}>
+                {isLogginActive && (
+                  <Login containerRef={ref => (this.current = ref)} />
+                )}
+                {!isLogginActive && (
+                  <Register containerRef={ref => (this.current = ref)} />
+                )}
+              </div>
+              <RightSide
+                current={current}
+                currentActive={currentActive}
+                containerRef={ref => (this.rightSide = ref)}
+                onClick={this.changeState.bind(this)}
+              />
+            </div>
+          </div>
+          <Footer />
+        </>
+        )
+    }
+}
+
+const RightSide = props => {
+    return (
+      <div
+        className="right-side"
+        ref={props.containerRef}
+        onClick={props.onClick}
+      >
+        <div className="inner-container">
+          <div className="text">{props.current}</div>
+        </div>
+      </div>
+    );
+    
+  };
+
+
